feat(auth): role and password encryption

+++ 4oModulo-FlaskAuth/app.py
@@ -2,6 +2,7 @@ from flask import Flask, request, jsonify
 from database import db
 from models.user import User
 from flask_login import LoginManager, login_user, current_user, logout_user, login_required
+import bcrypt
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = "uau"
@@ -32,7 +33,7 @@ def login():
   if username and password:
     user = User.query.filter_by(username=username).first()
 
-    if user and user.password == password:
+    if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
       login_user(user)
       print(current_user.is_authenticated)
       return jsonify({
@@ -60,7 +61,8 @@ def create_user():
   password = data.get('password')
 
   if username and password:
-     user = User(username=username, password=password)
+     hashed_pw = bcrypt.hashpw(str.encode(password),bcrypt.gensalt())
+     user = User(username=username, password=hashed_pw, role='user')
      db.session.add(user)
      db.session.commit()
      return jsonify({
@@ -69,7 +71,6 @@ def create_user():
   return jsonify({
     "mensagem":"Credenciais inválidas"
   }), 400
-
 
 
 @app.route("/user/<int:id>", methods=['GET'])
@@ -93,6 +94,12 @@ def read_user(id):
 def update_user(id):
   user = User.query.get(id)
   data = request.json
+
+  if id != current_user.id and current_user.role == "role":
+    return jsonify({
+    "mensagem": f"Operação não permitida"
+  }), 403 
+
   if user and data.get("password"):
     user.password = data.get("password")
     db.session.commit()
@@ -108,12 +115,18 @@ def update_user(id):
 
 @app.route("/user/<int:id>", methods=['DELETE'])
 @login_required
-def read_user(id):
+def delete_user(id):
   user = User.query.get(id)
+
+  if current_user.role != "admin":
+    return jsonify({
+    "mensagem": f"Operação não permitida"
+  }),403
 
   if user:
     if id == current_user.id:
-      logout()
+      logout_user()
+      
     db.session.delete(user)
     db.session.commit()
     return jsonify({
